Only need to call function to prepare redirection + norminette
