+++ src/error_syntax_pipes.c
@@ -1,5 +1,25 @@
 #include "../include/minishell.h"
 
+void	print_error_token(int check)
+{
+	if (check == ERROR_1PIPE)
+		ft_putstr_fd("bash: syntax error near unexpected token `|'\n", 2);
+	else if (check == ERROR_2PLUSPIPE)
+		ft_putstr_fd("bash: syntax error near unexpected token `||'\n", 2);
+	else if (check == ERROR_REDIRECTION_INPUT)
+		ft_putstr_fd("bash: syntax error near unexpected token `newline'\n", 2);
+	else if (check == ERROR_3REDIRECTION_INPUT)
+		ft_putstr_fd("bash: syntax error near unexpected token `>'\n", 2);
+	else if (check == ERROR_4PLUSREDIRECTION_INPUT)
+		ft_putstr_fd("bash: syntax error near unexpected token `>>'\n", 2);
+	else if (check == ERROR_REDIRECTION_OUTPUT)
+		ft_putstr_fd("bash: syntax error near unexpected token `newline'\n", 2);
+	else if (check == ERROR_4REDIRECTION_OUTPUT)
+		ft_putstr_fd("bash: syntax error near unexpected token `<'\n", 2);
+	else if (check == ERROR_5PLUSREDIRECTION_OUTPUT)
+		ft_putstr_fd("bash: syntax error near unexpected token `<<'\n", 2);
+}
+
 int	check_first(char *str)
 {
 	int	pipe_number;
@@ -26,10 +46,10 @@ int	check_first(char *str)
 
 int	check_last(char *str)
 {
-	char *temp;
-	int	pipe_number;
+	char	*temp;
+	int		pipe_number;
 
-	pipe_number = 0;	
+	pipe_number = 0;
 	temp = str + ft_strlen(str) - 1;
 	while (*temp == ' ' || *temp == '\t' || *temp == '\n')
 		temp--;
@@ -40,7 +60,7 @@ int	check_last(char *str)
 			while (*temp == '|')
 			{
 				pipe_number++;
-				temp--;;
+				temp--;
 			}
 			if (pipe_number == 1)
 				print_error_token(ERROR_1PIPE);
@@ -77,7 +97,7 @@ int	check_syntax_pipes(char *str, t_data *data, int check)
 {
 	if (check == 1)
 	{
-		if (check_first(str) == - 1 || check_last(str) == -1)
+		if (check_first(str) == -1 || check_last(str) == -1)
 		{
 			data->exitcode = 2;
 			return (ERROR);
